add label option to fuzzy rule functions
This change allows to get both numeric and labeled regions for
inputs/outputs. With this, you can get either a tuple of them three
points for membership function or a labeled that goes from Sn...S1, CE,
B1...Bn. This change will allow to code a function that cleans conflicts
in the rule base.