change generate fuzzy rule function
Generate fuzzy rule now expects one tuple for input and one for output,
containing a ndim list of values and a ndim list of region tuples. There
is also now a flag called only_regions that makes it so it does not
return the values on a rule generate, only fuzzy regions.