change clean_conflicting_rule_base to reflect other changes
Changed so it neglects the original data pair in equality check, yet
checks rule_base parameter for it in order to calculate degree. Also it
no longer uses the old rules based on list of dicts, it now recognizes
the current way of rules, a list of tuples. Still wonder why no rules
were removed in my tests.