change functions inputs type to dict
Change determine_degrees_and_assign, generate_fuzzy_rule and
generate_time_series_rule_base to accept as inputs a dict with values as
way to provide a mapping to their regions. This was done as a design
change due to the nature of fuzzy inference, since it also requires the
regions to be known aside from the rules.